chore: move secret fetch to aws module
