+++ scripts/arbitrum_harvest.py
@@ -6,7 +6,7 @@ from web3 import Web3
 from config.constants import MULTICHAIN_CONFIG
 from config.enums import Network
 from src.general_harvester import GeneralHarvester
-from src.utils import get_secret
+from src.aws import get_secret
 from src.web3_utils import get_strategies_from_registry
 
 logging.basicConfig(level=logging.INFO)
