+++ scripts/ftm_harvest.py
@@ -7,7 +7,7 @@ from config.constants import FTM_VAULTS
 from config.constants import MULTICHAIN_CONFIG
 from config.enums import Network
 from src.general_harvester import GeneralHarvester
-from src.utils import get_secret
+from src.aws import get_secret
 from src.web3_utils import get_strategy_from_vault
 
 logging.basicConfig(level=logging.INFO)
