+++ scripts/poly_harvest.py
@@ -9,7 +9,7 @@ from config.constants import POLY_OLD_STRATEGY_1
 from config.constants import POLY_OLD_STRATEGY_2
 from config.enums import Network
 from src.general_harvester import GeneralHarvester
-from src.utils import get_secret
+from src.aws import get_secret
 from src.web3_utils import get_strategies_from_registry
 
 logging.basicConfig(level=logging.INFO)
