+++ scripts/rebase.py
@@ -3,7 +3,7 @@ import logging
 from config.enums import Network
 from src.rebaser import Rebaser
 from src.utils import get_node_url
-from src.utils import get_secret
+from src.aws import get_secret
 
 logging.basicConfig(level=logging.INFO)
